refactor: move chain configs to a dedicated directory
